Beta (#5)
* - Added sample identifier in sample test.
- Autopopulate lab reference Id with timestamp
- add global properties for file extensions  
- add void test result 
- implement the freezing condition on lab test sample manage
- remove the comments field from lab test result form

* - Added The Resource Controller 
- change the sample_identifier varchar()value  from 50 to 255
- added sample identifier in manage sample 
- autopopulate lab reference with timestamp plus test_type.shortname
- autopopulate sample identifier with timestamp plu specimen type

* Added resource controllers

* updated Test Order REST

* updated test order rest

* Added Test request form

* - Complete UI and submition of test request form
- Also added the voided values check in api for test Result

* - Added the regex validation for float datatype and RegexDatatype  in
result form.

* - Added Hint field in labTestAttributeType 
- Resolved the reported bugs from QA

* completed unit tests except for fews which functionality was not
available or they were too complicated to do unit testing on them. Also
updated rest controllers and added test for each resource following
Openmrs Rest test pattern.

* change the fontawesome icon with images  in protlet

* add imagess

* -Handle the validation on Test Results
-Implemented the Length ,Range and Regex feature in
addLabTestAttributeTyp.

* Added grouping for metadata

* Final Code

* added the sub -multiset option within a group  feature 
added encounter type, date created   in Order view

* remove the group_id from the testAttributeType table

* add labTestAttributeType and Rest Resources

* update version 1.0.1

* update version 1.0.1 in archiver/pom.properties

* Remove comments and completed the TODO tasks

+++ omod/src/main/java/org/openmrs/module/commonlabtest/web/controller/LabTestAttributeTypeController.java
@@ -2,6 +2,7 @@ package org.openmrs.module.commonlabtest.web.controller;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
+import org.openmrs.api.context.Context;
 import org.openmrs.customdatatype.CustomDatatypeUtil;
 import org.openmrs.module.commonlabtest.LabTestAttribute;
 import org.openmrs.module.commonlabtest.LabTestAttributeType;
@@ -47,6 +48,7 @@ public class LabTestAttributeTypeController {
 	@RequestMapping(method = RequestMethod.GET, value = "/module/commonlabtest/addLabTestAttributeType.form")
 	public String showForm(ModelMap model, @RequestParam(value = "error", required = false) String error,
 	        @RequestParam(value = "uuid", required = false) String uuid) {
+		
 		LabTestAttributeType attributeType;
 		if (uuid == null || uuid.equalsIgnoreCase("")) {
 			attributeType = new LabTestAttributeType();
@@ -54,7 +56,6 @@ public class LabTestAttributeTypeController {
 			attributeType = commonLabTestService.getLabTestAttributeTypeByUuid(uuid);
 			List<LabTestAttribute> labTestAttributes = commonLabTestService.getLabTestAttributes(attributeType,
 			    Boolean.FALSE);
-			
 			if (labTestAttributes.size() > 0) {
 				model.addAttribute("available", Boolean.TRUE);
 			} else {
@@ -71,6 +72,7 @@ public class LabTestAttributeTypeController {
 	public String onSubmit(ModelMap model, HttpSession httpSession,
 	        @ModelAttribute("anyRequestObject") Object anyRequestObject, HttpServletRequest request,
 	        @ModelAttribute("attributeType") LabTestAttributeType attributeType, BindingResult result) {
+		
 		String status = "";
 		try {
 			if (result.hasErrors()) {
@@ -84,18 +86,16 @@ public class LabTestAttributeTypeController {
 			} else {
 				
 				commonLabTestService.saveLabTestAttributeType(attributeType);
-				StringBuilder sb = new StringBuilder();
-				sb.append("Lab Test Attribute with Uuid :");
-				sb.append(attributeType.getUuid());
-				sb.append(" is  saved!");
-				status = sb.toString();
+				StringBuilder subString = new StringBuilder();
+				subString.append("Lab Test Attribute with Uuid :");
+				subString.append(attributeType.getUuid());
+				subString.append(" is  saved!");
+				status = subString.toString();
 			}
 		}
 		catch (Exception e) {
-			status = "Error! could not save Lab Test Attribute Type";
-			//status = e.getLocalizedMessage();
+			status = "could not save Lab Test Attribute Type";
 			e.printStackTrace();
-			
 			model.addAttribute("error", status);
 			if (attributeType.getLabTestAttributeTypeId() == null) {
 				return "redirect:addLabTestAttributeType.form";
@@ -111,18 +111,26 @@ public class LabTestAttributeTypeController {
 	@RequestMapping(method = RequestMethod.POST, value = "/module/commonlabtest/retirelabtestattributetype.form")
 	public String onRetire(ModelMap model, HttpSession httpSession, HttpServletRequest request,
 	        @RequestParam("uuid") String uuid, @RequestParam("retireReason") String retireReason) {
-		LabTestAttributeType attributeType = commonLabTestService.getLabTestAttributeTypeByUuid(uuid);
-		String status;
+		
+		LabTestAttributeType attributeType = Context.getService(CommonLabTestService.class).getLabTestAttributeTypeByUuid(
+		    uuid);
+		String status = "";
+		if (Context.getAuthenticatedUser() == null) {
+			return "redirect:../../login.htm";
+		}
+		if (Context.getAuthenticatedUser() == null) {
+			return "redirect:../../login.htm";
+		}
 		try {
 			commonLabTestService.retireLabTestAttributeType(attributeType, retireReason);
-			StringBuilder sb = new StringBuilder();
-			sb.append("Lab Test Attribute with Uuid :");
-			sb.append(attributeType.getUuid());
-			sb.append(" is  retired!");
-			status = sb.toString();
+			StringBuilder subString = new StringBuilder();
+			subString.append("Lab Test Attribute with Uuid :");
+			subString.append(attributeType.getUuid());
+			subString.append(" is  retired!");
+			status = subString.toString();
 		}
 		catch (Exception e) {
-			status = "Error! could not save Lab Test Attribute Type";
+			status = "could not retire Lab Test Attribute Type";
 			e.printStackTrace();
 			model.addAttribute("error", status);
 			if (attributeType.getLabTestAttributeTypeId() == null) {
@@ -139,19 +147,23 @@ public class LabTestAttributeTypeController {
 	@RequestMapping(method = RequestMethod.POST, value = "/module/commonlabtest/deletelabtestattributetype.form")
 	public String onDelete(ModelMap model, HttpSession httpSession, HttpServletRequest request,
 	        @RequestParam("uuid") String uuid) {
-		LabTestAttributeType attributeType = commonLabTestService.getLabTestAttributeTypeByUuid(uuid);
+		LabTestAttributeType attributeType = Context.getService(CommonLabTestService.class).getLabTestAttributeTypeByUuid(
+		    uuid);
 		String status;
+		if (Context.getAuthenticatedUser() == null) {
+			return "redirect:../../login.htm";
+		}
 		try {
 			commonLabTestService.deleteLabTestAttributeType(attributeType, true);
-			StringBuilder sb = new StringBuilder();
-			sb.append("Lab Test Attribute with Uuid :");
-			sb.append(attributeType.getUuid());
-			sb.append(" is permanently deleted!");
-			status = sb.toString();
+			StringBuilder subString = new StringBuilder();
+			subString.append("Lab Test Attribute with Uuid :");
+			subString.append(attributeType.getUuid());
+			subString.append(" is permanently deleted!");
+			status = subString.toString();
 		}
 		catch (Exception exception) {
-			//status = exception.getLocalizedMessage();
-			status = "Error! could not save Lab Test Attribute Type";
+			// status = exception.getLocalizedMessage();
+			status = "could not delete Lab Test Attribute Type";
 			exception.printStackTrace();
 			model.addAttribute("error", status);
 			return "redirect:addLabTestAttributeType.form?uuid=" + attributeType.getUuid();
